% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptBayes-package.R
\docType{package}
\name{adaptBayes-package}
\alias{adaptBayes-package}
\alias{adaptBayes}
\title{The 'adaptBayes' package.}
\description{
R package for Boonstra, Philip S. and Barbaro, Ryan P., “Incorporating Historical Models with Adaptive Bayesian Updates” (2018) In Press at Biostatistics
}
\references{
R package for Boonstra, Philip S. and Barbaro, Ryan P., “Incorporating Historical Models with Adaptive Bayesian Updates” (2018) In Press at Biostatistics

Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
