% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{count_stan_divergences}
\alias{count_stan_divergences}
\title{Count the stan divergences}
\usage{
count_stan_divergences(stan_fit)
}
\arguments{
\item{stan_fit}{an R object of class stanfit}
}
\value{
The number of divergent transitions across all chains
}
\description{
Helper function to count the number of divergent transitions in a stanfit
object
}
