% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_sab.R
\name{glm_sab}
\alias{glm_sab}
\title{Fit GLM with the 'sensible adaptive bayes' prior}
\usage{
glm_sab(stan_fit = NA, stan_path, y = c(0, 1),
  x_standardized = matrix(0, length(y), 6), alpha_prior_mean = rep(0,
  3), alpha_prior_cov = diag(1, 3), aug_projection = diag(1, 3),
  phi_mean = 0.5, phi_sd = 2.5, beta_orig_scale = 1,
  beta_aug_scale = 1, local_dof = 1, global_dof = 1,
  slab_precision = (1/15)^2, only_prior = F, mc_warmup = 50,
  mc_iter_after_warmup = 50, mc_chains = 1, mc_thin = 1,
  mc_stepsize = 0.1, mc_adapt_delta = 0.9, mc_max_treedepth = 15,
  ntries = 1, eigendecomp_hist_var = NULL,
  scale_to_variance225 = NULL)
}
\arguments{
\item{stan_fit}{an R object of class stanfit, which allows the function to run
without recompiling the stan code.}

\item{stan_path}{(character) a path pointing to a .stan file, which indicates
the stan code to compile and run. If both stan_fit and stan_path are provided,
stan_fit takes precedence.}

\item{y}{(vector) outcomes corresponding to the type of glm desired. This should
match whatever datatype is expected by the stan program.}

\item{x_standardized}{(matrix) matrix of numeric values with number of rows equal
to the length of y and number of columns equal to p+q. It is assumed without
verification that each column is standardized to whatever scale the prior
expects - in Boonstra and Barbaro, all predictors are marginally generated to have
mean zero and unit variance, so no standardization is conducted. In practice,
all data should be standardized to have a common scale before model fitting.
If regression coefficients on the natural scale are desired, they be easily obtained
through unstandardizing.}

\item{alpha_prior_mean}{(vector) p-length vector giving the mean of alpha from the
historical analysis, corresponds to m_alpha in Boonstra and Barbaro}

\item{alpha_prior_cov}{(matrix) pxp positive definite matrix giving the variance of
alpha from the historical analysis, corresponds to S_alpha in Boonstra and Barbaro}

\item{phi_mean}{(real) mean of phi corresponding to a normal distribution, support is truncated to \link{0,1}}

\item{phi_sd}{(pos. real) sd of phi corresponding to a normal distribution, support is truncated to \link{0,1}}

\item{beta_orig_scale}{(pos. real) constants indicating the prior scale of the
horseshoe. Both values correspond to 'c' in the notation of Boonstra and Barbaro,
because that paper never considers beta_orig_scale!=beta_aug_scale}

\item{beta_aug_scale}{(pos. real) constants indicating the prior scale of the
horseshoe. Both values correspond to 'c' in the notation of Boonstra and Barbaro,
because that paper never considers beta_orig_scale!=beta_aug_scale}

\item{local_dof}{(pos. integer) numbers indicating the degrees of freedom for
lambda_j and tau, respectively. Boonstra, et al. never considered local_dof != 1
or global_dof != 1.}

\item{global_dof}{(pos. integer) numbers indicating the degrees of freedom for
lambda_j and tau, respectively. Boonstra, et al. never considered local_dof != 1
or global_dof != 1.}

\item{slab_precision}{(pos. real) the slab-part of the regularized horseshoe,
this is equivalent to (1/d)^2 in the notation of Boonstra and Barbaro}

\item{only_prior}{(logical) should all data be ignored, sampling only from the prior?}

\item{mc_warmup}{number of MCMC warm-up iterations}

\item{mc_iter_after_warmup}{number of MCMC iterations after warm-up}

\item{mc_chains}{number of MCMC chains}

\item{mc_thin}{every nth draw to keep}

\item{mc_stepsize}{positive stepsize}

\item{mc_adapt_delta}{between 0 and 1}

\item{mc_max_treedepth}{max tree depth}

\item{ntries}{(pos. integer) the stan function will run up to this many times,
stopping either when the number of divergent transitions* is zero or when ntries
has been reached. The reported fit will be that with the fewest number of divergent iterations.}

\item{eigendecomp_hist_var}{fill}

\item{scale_to_variance225}{fill}
}
\value{
\code{list} object containing the draws and other information.
}
\description{
Program for fitting a GLM equipped with the 'sensible adaptive bayes' prior
evaluated in the manuscript.
}
