% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_studt.R
\name{glm_studt}
\alias{glm_studt}
\title{Fit GLM with a regularized student-t prior regression coefficients}
\usage{
glm_studt(stan_fit = stanmodels$RegStudT, stan_path, y = c(0, 1),
  x_standardized = matrix(0, length(y), 3), beta_scale = 1, dof = 1,
  slab_precision = (1/15)^2, only_prior = F, mc_warmup = 50,
  mc_iter_after_warmup = 50, mc_chains = 1, mc_thin = 1,
  mc_stepsize = 0.1, mc_adapt_delta = 0.9, mc_max_treedepth = 15,
  ntries = 1)
}
\arguments{
\item{stan_fit}{an R object of class stanfit, which allows the function to run
without recompiling the stan code.}

\item{stan_path}{(character) a path pointing to a .stan file, which indicates
the stan code to compile and run. If both stan_fit and stan_path are provided,
stan_fit takes precedence.}

\item{y}{(vector) outcomes corresponding to the type of glm desired. This should
match whatever datatype is expected by the stan program.}

\item{x_standardized}{(matrix) matrix of numeric values with number of rows equal
to the length of y and number of columns equal to p+q. It is assumed without
verification that each column is standardized to whatever scale the prior
expects - in Boonstra and Barbaro, all predictors are marginally generated to have
mean zero and unit variance, so no standardization is conducted. In practice,
all data should be standardized to have a common scale before model fitting.
If regression coefficients on the natural scale are desired, they be easily obtained
through unstandardizing.}

\item{beta_scale}{(pos. real) constants indicating the prior scale of the student-t prior.}

\item{dof}{(pos. integer) degrees of freedom for the student-t prior}

\item{slab_precision}{(pos. real) the slab-part of the regularized horseshoe,
this is equivalent to (1/d)^2 in the notation of Boonstra and Barbaro}

\item{only_prior}{(logical) should all data be ignored, sampling only from the prior?}

\item{mc_warmup}{number of MCMC warm-up iterations}

\item{mc_iter_after_warmup}{number of MCMC iterations after warm-up}

\item{mc_chains}{number of MCMC chains}

\item{mc_thin}{every nth draw to keep}

\item{mc_stepsize}{positive stepsize}

\item{mc_adapt_delta}{between 0 and 1}

\item{mc_max_treedepth}{max tree depth}

\item{ntries}{(pos. integer) the stan function will run up to this many times,
stopping either when the number of divergent transitions* is zero or when ntries
has been reached. The reported fit will be that with the fewest number of divergent iterations.}
}
\value{
\code{list} object containing the draws and other information.
}
\description{
Program for fitting a GLM equipped with a regularized student-t prior on the regression
coefficients, parametrized using the normal-inverse-gamma distribution. The 'regularization'
refers to the fact that the inverse-gamma scale is has a finite upper bound that
it smoothly approaches. This method was not used in the simulation study but was
used in the data analysis. Specifically, it corresponds to 'PedRESC2'.
}
